requirements.txt
alchemical==1.0.2
alembic==1.13.1
apifairy==1.4.0
apispec==6.6.1
bidict==0.23.1
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
eventlet==0.36.1
Flask==3.0.2
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
Flask-Mail==0.9.1
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
flask-mongoengine @ git+https://github.com/idoshr/flask-mongoengine.git@e244408acf440c4208f7ddcd6e5d819cb472e4da
flask-sock==0.7.0
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-jsonschema==0.13.0
marshmallow-sqlalchemy==1.0.0
mongoengine==0.28.2
packaging==24.1
PyJWT==2.8.0
pymongo==4.7.3
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
SQLAlchemy==2.0.30
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
webargs==8.4.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
zope.event==5.0
zope.interface==7.0.3
mongoengine_goodjson